[refactor] ExecutorConfig 메소드 통합
